+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 ---------
 
+0.8.1 (2014-04-XX)
+++++++++++++++++++
+
+- Fixed: `Issue #27 <https://github.com/maxtepkeev/python-redmine/issues/27>`__ (Project and
+  Issue resources ``parent`` attribute was returned as a dict instead of being converted to
+  Resource object)
+
 0.8.0 (2014-03-27)
 ++++++++++++++++++